Confine secret minting to create-schnorr-account
The deploy scripts minted a secret when the env var was absent and
echoed it after spending real network time, an easy way to end up with
an unsaved key that controls a live registry. Now a single derive-only
script mints and prints the secret, and every other script requires
ZKPASSPORT_DEPLOYER_SECRET from the environment.
