+++ packages/aztec/scripts/package.json
@@ -5,8 +5,8 @@
   "scripts": {
     "e2e": "tsx e2e/run-e2e.ts",
     "format": "prettier -c -w e2e deployment",
-    "deployer:testnet": "tsx deployment/create-deployer.ts testnet",
-    "deployer:mainnet": "tsx deployment/create-deployer.ts mainnet",
+    "account:testnet": "tsx deployment/create-schnorr-account.ts testnet",
+    "account:mainnet": "tsx deployment/create-schnorr-account.ts mainnet",
     "deploy:testnet": "tsx deployment/deploy-testnet.ts",
     "deploy:mainnet": "tsx deployment/deploy-mainnet.ts",
     "seed:testnet": "tsx deployment/seed-registry.ts testnet",
